+++ src/main/java/dev/triangulation/graphics/DrawableLine.java
@@ -12,10 +12,10 @@ public class DrawableLine {
     }
 
     public void draw(Graphics graphics) {
-        graphics.drawLine(point1.getX(), point1.getY(), point2.getX(), point2.getY());
+        graphics.drawLine((int)point1.getX(), (int)point1.getY(), (int)point2.getX(), (int)point2.getY());
     }
 
     public static void drawLine(DrawablePoint point1, DrawablePoint point2, Graphics graphics){
-        graphics.drawLine(point1.getX(), point1.getY(), point2.getX(), point2.getY());
+        graphics.drawLine((int)point1.getX(), (int)point1.getY(), (int)point2.getX(), (int)point2.getY());
     }
 }
